+++ Changes.md
@@ -27,6 +27,7 @@ and BSpline curves to their endpoints automatically, without manual management o
 - OSLCode : The OSL shader is now compiled on demand, rather than every time the node is edited. This avoids many redundant attempts at recompilation when loading nodes with many parameters.
 - ArnoldMeshLight, RenderManMeshLight : Added support for deformation motion blur (#6869). In the case of RenderMan, this only affects the camera-visible mesh, since RenderMan doesn't yet support deformation for the light itself.
 - TweakPlug : Added `SetExpressionInclude` and `SetExpressionExclude` modes for tweaking set expressions.
+- Graph Editor : Added location bar for navigation through Boxes and References. Text and button interactions can be used to navigate the node hierarchy.
 
 Fixes
 -----
@@ -62,6 +63,8 @@ API
 - MeshLight : Added based class to simplify the implementation of renderer-specific mesh light nodes.
 - PathColumn : `headerData()` is now passed the root Path.
 - SetExpressionAlgo : Added new namespace with functions for evaluating and editing set expressions.
+- GraphComponentPath : Added `setFromComponent()`
+- BreadCrumbsWidget : Added widget for interacting with paths using a combination of button widgets and text entry.
 
 Breaking Changes
 ----------------